Move code to extension file
