+++ ComicRatingTests/GeneratedMocks.swift
@@ -1,4 +1,4 @@
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Configurator/ComicHistoryInitializer.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Configurator/ComicHistoryInitializer.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicHistoryComicHistoryInitializer.swift
@@ -183,7 +183,7 @@ class ComicHistoryModuleBuilderStub: ComicHistoryModuleBuilder {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Interactor/ComicHistoryInteractorInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Interactor/ComicHistoryInteractorInput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicHistoryComicHistoryInteractorInput.swift
@@ -299,7 +299,7 @@ class ComicHistoryInteractorInputStub: ComicHistoryInteractorInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Interactor/ComicHistoryInteractorOutput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/Interactor/ComicHistoryInteractorOutput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicHistoryComicHistoryInteractorOutput.swift
@@ -448,7 +448,7 @@ class ComicHistoryInteractorOutputStub: ComicHistoryInteractorOutput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewController.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewController.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicHistoryViewController.swift
@@ -630,7 +630,7 @@ class ComicHistoryViewControllerStub: ComicHistoryViewController {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewInput.swift at 2020-08-21 13:40:21 +0000
 
 @testable import ComicRating
 //
@@ -804,7 +804,7 @@ class ComicHistoryViewInputStub: ComicHistoryViewInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewOutput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ComicHistory/View/ComicHistoryViewOutput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicHistoryComicHistoryViewOutput.swift
@@ -882,7 +882,7 @@ class ComicHistoryViewOutputStub: ComicHistoryViewOutput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Configurator/ShowComicInitializer.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Configurator/ShowComicInitializer.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ShowComicInitializer.swift
@@ -1067,7 +1067,7 @@ class ShowComicModuleBuilderStub: ShowComicModuleBuilder {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Interactor/ShowComicInteractorInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Interactor/ShowComicInteractorInput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ShowComicShowComicInteractorInput.swift
@@ -1202,7 +1202,7 @@ class ShowComicInteractorInputStub: ShowComicInteractorInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Interactor/ShowComicInteractorOutput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/Interactor/ShowComicInteractorOutput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ShowComicShowComicInteractorOutput.swift
@@ -1339,7 +1339,7 @@ class ShowComicInteractorOutputStub: ShowComicInteractorOutput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewController.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewController.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ShowComicShowComicViewController.swift
@@ -1857,7 +1857,7 @@ class ShowComicViewControllerStub: ShowComicViewController {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewInput.swift at 2020-08-21 13:40:21 +0000
 
 @testable import ComicRating
 //
@@ -2111,7 +2111,7 @@ class ShowComicViewInputStub: ShowComicViewInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewOutput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/ShowComic/View/ShowComicViewOutput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ShowComicShowComicViewOutput.swift
@@ -2244,7 +2244,7 @@ class ShowComicViewOutputStub: ShowComicViewOutput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/Interactor/TabBarInteractorInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/Interactor/TabBarInteractorInput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  TabBarTabBarInteractorInput.swift
@@ -2297,7 +2297,7 @@ class MockTabBarInteractorInput: TabBarInteractorInput, Cuckoo.ProtocolMock {
 
 class TabBarInteractorInputStub: TabBarInteractorInput {}
 
-// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/Router/TabBarRouterInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/Router/TabBarRouterInput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  TabBarRouterInput.swift
@@ -2376,7 +2376,7 @@ class TabBarRouterInputStub: TabBarRouterInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/View/TabBarViewInput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/View/TabBarViewInput.swift at 2020-08-21 13:40:21 +0000
 
 @testable import ComicRating
 //
@@ -2460,7 +2460,7 @@ class TabBarViewInputStub: TabBarViewInput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/View/TabBarViewOutput.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Modules/TabBar/View/TabBarViewOutput.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  TabBarTabBarViewOutput.swift
@@ -2538,7 +2538,7 @@ class TabBarViewOutputStub: TabBarViewOutput {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Api/ComicApiService.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Api/ComicApiService.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicApiService.swift
@@ -2788,7 +2788,7 @@ class ComicApiServiceStub: ComicApiService {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Api/ComicApiServiceInterface.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Api/ComicApiServiceInterface.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicApiClient.swift
@@ -3000,7 +3000,7 @@ class ComicApiServiceInterfaceStub: ComicApiServiceInterface {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Api/ImageDownloaderService.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Api/ImageDownloaderService.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ImageDownloader.swift
@@ -3084,7 +3084,7 @@ class ImageDownloaderServiceStub: ImageDownloaderService {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Api/ImageDownloaderServiceInterface.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Api/ImageDownloaderServiceInterface.swift at 2020-08-21 13:40:21 +0000
 
 @testable import ComicRating
 //
@@ -3165,7 +3165,7 @@ class ImageDownloaderServiceInterfaceStub: ImageDownloaderServiceInterface {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Storage/ComicStorageService.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Storage/ComicStorageService.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicStorageService.swift
@@ -3275,7 +3275,7 @@ class ComicStorageServiceStub: ComicStorageService {
     }
 }
 
-// MARK: - Mocks generated from file: ComicRating/Services/Storage/ComicStorageServiceInterface.swift at 2020-08-21 13:07:40 +0000
+// MARK: - Mocks generated from file: ComicRating/Services/Storage/ComicStorageServiceInterface.swift at 2020-08-21 13:40:21 +0000
 
 //
 //  ComicStorageServiceInterface.swift
